pyproject.toml
[tool.poetry]
name = "webpage"
version = "0.1.0"
description = "elecciones"
authors = ["Grace nickol"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.7"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
django-extensions = "^3.1.1"
django-debug-toolbar = "^3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
